+++ core/cantera/base/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class CanteraBaseConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'core.cantera.base'
+    label = 'base_cantera'
+
