+++ core/cantera/listar_cantera/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ListarCanteraConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'core.cantera.listar_cantera'
+    label = 'listar_cantera'
